Add test zero

+++ src/libs/square_root.cpp
@@ -1,6 +1,6 @@
 #include "square_root.h"
 
-double square_root(double num)
+long double square_root(long double num)
 {
     return sqrt(num);
 }
